pipeline/requirements.txt
pandas==2.2.3
numpy==1.26.4
pyarrow==17.0.0
tqdm==4.66.5
mariadb==1.1.10
Pillow==10.4.0
sentence-transformers==3.1.1
python-dotenv==1.0.1
